pyproject.toml
[project]
name = "boundary-analyzer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0",
  "pandas>=2.2.0",
  "PyYAML>=6.0.1",
  "dash>=2.14.0",
  "plotly>=5.18.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
